Use CancelProfile to set a profile to Cancelled
The ProfileCloser object will be in charge of everything having to do
with a profile being closed. This will eventually entail:

* Set cancelled to true on the Profile object
* Removing the profile picture
* Prepend CANCELLED to email address
* Change name to CANCELLED ACCOUNT
* Tracking the specified reason they closed the account

Right now it only does "Set cancelled to true". This uses the
`time_for_a_boolean` gem which expect a column to be a `datetime` and
be in the format <something>_at. In this case, the column is
`closed_at`. You can interact with that column as `closed` like it's a
boolean column in most cases. When set to "true" it records the current
time. This is very useful for analytics later.

Read more about time_for_a_boolean here:
https://github.com/calebthompson/time_for_a_boolean