Prepend email with cancelled when canceling account
This prepends "cancelled" to the users email address when the account
is cancelled. This is so we don't accidentally email them.

An issue came up when trying to make this work. Devise can be
configured to confirm email addresses, both on user sign up and if a
user changes their email. This was the case with this app. So when I
tried to set the email to `cancelleduser@example.com` it triggered the
reconfirmation process. No good.

To get around this, I called this method on user before changing their
email:
https://github.com/plataformatec/devise/blob/master/lib/devise/models/co
nfirmable.rb#L162-L164 Seems to work well.