Redirect to homepage is profile is closed
Users should not be able to look at closed profiles, so it redirects
them to the homepage instead. Note the `@profile.profile` bit in
controller... I have no idea why a User is being located for the
Profiles controller but being called a profile... Somewhat concerning,
but I don't feel confident enough yet to change that code.