Fix regression where closed reason would be blank
The way profile.closed_reason is set is when you type in the "other"
box, it transfers that text over to the radio button's value. When the
forms submits, that radio button gets passed as the closed reason in
the params and trickles down to set the profile in ProfileCloser.

The ID of the radio button element ended with an underscore:
`#profile_closed_reason_` because the tag_value was set to an empty
string. It is now sets to `#profile_closed_reason_other` which is
similar to the other radio buttons