Search does not return closed users
On the homepage, when a user searches with an interest and/or zipcode,
it returns a list of (among other things) users. Before this change, it
was returning users with closed profiles. Now, the search does not.

`Profile.terms` seems responsible for this search (or at least the
person part of it) so I added a condition to the query `AND closed_at
IS NULL` so when it returns people/profiles/users, they are currently
"active" and not "closed".