Add Profile decorator with display_name
This will be used for all modified bits of data that relate to or come
from a profile. For example, the display_name method gets the first and
last name from the passed in profile and returns Firstname Lastinitial.

Simple delegator will delegate all methods that are normally on
profile. So `profile.first_name` will be the same as
`delegated_profile.first_name`