Replace instances of last name with last initial
The purpose of this change is to add privacy for users so no one can
see their last name.

The ProfileDecorator does a lot of the work here. There were only a few
places it made sense to decorate the object in the controller. The rest
of the time I instantiated the decorator in the view.

There are many instances of an IM window. These are `<div>` tags with
`recipient_lastname` attributes. These set the display name when you
"Send <user> a message"